ユーザー取得を共通化対応
